* Resources/test-player.png:   New default image for test player.
* Source/Engine/Image.h:
* Source/Game/World.cpp:
* Source/Engine/Engine.h:
* Source/Engine/Sprite.h:
* Source/Game/TestMode.cpp:
* Source/Game/GameObject.h:
* Source/Engine/Position.h:
* Source/Engine/Sprite.cpp:
* Source/Engine/Animation.h:
* Source/Engine/Rectangle.h:
* Source/Game/GameObject.cpp:
* Source/Engine/Position.cpp:
* Source/Engine/Rectangle.cpp:
* Source/Engine/Animation.cpp:
* Source/Game/Objects/Player.h:
* Source/Engine/SDL/SDLImage.h:
* Source/Engine/SDL/SDLEngine.h:
* Source/Engine/AnimationFrame.h:
* Source/Engine/SDL/SDLImage.cpp:
* Source/Game/Objects/Player.cpp:
* Source/Engine/SDL/SDLEngine.cpp:
* Source/Engine/AnimationFrame.cpp:
* Source/Game/ObjectBehaviors/DefaultCollidingBehavior.cpp:
  Add support for multiple collision rectangles for GameObjects.
Implement sprites and animations.

* BlockWorld.cproj:
  Added SDL_image as dependency and added new files for sprites and
  animations.

+++ Source/Engine/Position.cpp
@@ -7,7 +7,7 @@ namespace BadEngine {
 	{
 	}
 	
-	Position::Position(int x, int y) :
+	Position::Position(double x, double y) :
 		m_x(x),
 		m_y(y)
 	{
@@ -17,22 +17,22 @@ namespace BadEngine {
 	{
 	}
 	
-	void Position::setX(int x)
+	void Position::setX(double x)
 	{
 		m_x = x;
 	}
 	
-	void Position::setY(int y)
+	void Position::setY(double y)
 	{
 		m_y = y;
 	}
 	
-	int Position::getX()
+	double Position::getX()
 	{
 		return m_x;
 	}
 	
-	int Position::getY()
+	double Position::getY()
 	{
 		return m_y;
 	}
